+++ planner/core/casetest/BUILD.bazel
@@ -5,7 +5,6 @@ go_test(
     timeout = "moderate",
     srcs = [
         "expression_rewriter_test.go",
-        "integration_partition_test.go",
         "integration_test.go",
         "main_test.go",
         "partition_pruner_test.go",
@@ -14,7 +13,6 @@ go_test(
         "predicate_simplification_test.go",
         "stats_test.go",
         "tiflash_selection_late_materialization_test.go",
-        "window_push_down_test.go",
     ],
     data = glob(["testdata/**"]),
     flaky = True,
