+++ planner/core/casetest/partition/testdata/partition_pruner_in.json
