+++ planner/core/casetest/partition/testdata/partition_pruner_out.json
